+++ src/test/java/blackjack/domain/participant/ParticipantTest.java
@@ -2,7 +2,6 @@ package blackjack.domain.participant;
 
 import blackjack.domain.Deck;
 import blackjack.domain.HandGenerator;
-import blackjack.domain.Name;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Number;
 import blackjack.domain.card.Shape;
